+++ src/SpeakEasy/HttpClientSettings.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using SpeakEasy.ArrayFormatters;
 using SpeakEasy.Authenticators;
-using SpeakEasy.Instrumentation;
+using SpeakEasy.Middleware;
 using SpeakEasy.Serializers;
 
 namespace SpeakEasy
@@ -20,19 +20,12 @@ namespace SpeakEasy
         {
             Serializers = new List<ISerializer>();
             Authenticator = new NullAuthenticator();
-            InstrumentationSink = new NullInstrumentationSink();
             NamingConvention = new DefaultNamingConvention();
-            UserAgent = SpeakEasy.UserAgent.SpeakEasy;
             ArrayFormatter = new MultipleValuesArrayFormatter();
 
             Serializers.Add(new DefaultJsonSerializer());
             Serializers.Add(new TextPlainSerializer());
         }
-
-        /// <summary>
-        /// The logging mechanism the client will use
-        /// </summary>
-        public IInstrumentationSink InstrumentationSink { get; set; }
 
         /// <summary>
         /// Any custom authentication required to access the http api
@@ -45,9 +38,12 @@ namespace SpeakEasy
         public List<ISerializer> Serializers { get; set; }
 
         /// <summary>
-        /// The user agent the web client will send when making http requests
+        /// The available serialiazers
         /// </summary>
-        public IUserAgent UserAgent { get; set; }
+        internal List<IHttpMiddleware> Middleware { get; } = new List<IHttpMiddleware>
+        {
+            new UserAgentMiddleware()
+        };
 
         /// <summary>
         /// The array formatter that will be used to format query string array paramters
@@ -106,6 +102,44 @@ namespace SpeakEasy
             }
 
             throw new InvalidOperationException("The http client settings are not valid.");
+        }
+
+        public int MiddlewareCount => Middleware.Count;
+
+        public void AppendMiddleware(IHttpMiddleware middleware)
+        {
+            Middleware.Add(middleware);
+        }
+
+        public void PrependMiddleware(IHttpMiddleware middleware)
+        {
+            Middleware.Insert(0, middleware);
+        }
+
+        public bool HasMiddleware<TMiddleware>()
+            where TMiddleware : IHttpMiddleware
+        {
+            return Middleware.Any(t => t is TMiddleware);
+        }
+
+        public void ReplaceMiddleware<TMiddleware>(TMiddleware replacement)
+            where TMiddleware : IHttpMiddleware
+        {
+            var index = RemoveMiddleware<TMiddleware>();
+            Middleware.Insert(index, replacement);
+        }
+
+        public int RemoveMiddleware<TMiddleware>()
+            where TMiddleware : IHttpMiddleware
+        {
+            if (!HasMiddleware<TMiddleware>())
+            {
+                throw new ArgumentException();
+            }
+
+            var index = Middleware.FindIndex(t => t is TMiddleware);
+            Middleware.RemoveAt(index);
+            return index;
         }
     }
 }