+++ src/SpeakEasy/IHttpClient.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -11,24 +10,9 @@ namespace SpeakEasy
     public interface IHttpClient
     {
         /// <summary>
-        /// An event that is raised before each request is run
-        /// </summary>
-        event EventHandler<BeforeRequestEventArgs> BeforeRequest;
-
-        /// <summary>
-        /// An event that is raised after each request is s
-        /// </summary>
-        event EventHandler<AfterRequestEventArgs> AfterRequest;
-
-        /// <summary>
         /// The root resource for this http client, all calls will be relative to this resource
         /// </summary>
         Resource Root { get; }
-
-        /// <summary>
-        /// The currently set user agent
-        /// </summary>
-        IUserAgent UserAgent { get; }
 
         /// <summary>
         /// Executes an http get request asynchronously